+++ FashopBackend.Domain/Services/ProductService.cs
@@ -35,7 +35,7 @@ namespace FashopBackend.Core.Services
             return product;
         }
 
-        public Product Edit(int inputId, string inputName, string description, double price, int sale, Brand brand,
+        public Product Edit(int inputId, string inputName, string description, decimal price, int sale, Brand brand,
             IEnumerable<Category> categories, IEnumerable<ProductImage> productImages)
         {
             Product product = _productRepository.GetIncluded(inputId);
