+++ FashopBackend/Graphql/Products/AddProductInput.cs
@@ -2,5 +2,5 @@
 
 namespace FashopBackend.Graphql.Products
 {
-    public record AddProductInput(string Name, string Descriptions, double Price, int Sale, int BrandId, IEnumerable<int> CategoryIds, IEnumerable<string> ImageUrls);
+    public record AddProductInput(string Name, string Descriptions, decimal Price, int Sale, int BrandId, IEnumerable<int> CategoryIds, IEnumerable<string> ImageUrls);
 }
